+++ b_lambda_layer_common_test/integration/app.py
@@ -1,6 +1,7 @@
 from aws_cdk.core import App
-from b_lambda_layer_common_test.integration.infrastructure import Infrastructure
+
+from b_lambda_layer_common_test.integration.infrastructure.main_stack import MainStack
 
 app = App()
-Infrastructure(app)
+MainStack(app)
 app.synth()