+++ b_lambda_layer_common_test/integration/tests/test_layer_v2.py
@@ -6,7 +6,7 @@ from botocore.response import StreamingBody
 from b_lambda_layer_common_test.integration.infrastructure.main_stack import MainStack
 
 
-def test_RESOURCE_lambda_layer_WITH_deployed_lambda_function_1_EXPECT_execution_successful():
+def test_RESOURCE_lambda_layer_v2_WITH_deployed_lambda_function_3_EXPECT_execution_successful():
     """
     Test whether the layer provides necessary functionality.
 
@@ -17,7 +17,7 @@ def test_RESOURCE_lambda_layer_WITH_deployed_lambda_function_1_EXPECT_execution_
 
     # Invoke specific lambda function.
     response = lambda_client.invoke(
-        FunctionName=MainStack.get_output(MainStack.LAMBDA_FUNCTION_1_NAME_KEY),
+        FunctionName=MainStack.get_output(MainStack.LAMBDA_FUNCTION_3_NAME_KEY),
         InvocationType='RequestResponse'
     )
 
@@ -27,5 +27,5 @@ def test_RESOURCE_lambda_layer_WITH_deployed_lambda_function_1_EXPECT_execution_
     data = json.loads(''.join(data))
 
     # Assert that the result is as expected.
-    assert data['Boto3Version'] == '1.16.35', data
-    assert data['BotocoreVersion'] == '1.19.35', data
+    assert data['Boto3Version'] is not None, data
+    assert data['BotocoreVersion'] is not None, data